+++ components/suggest/src/schema.rs
@@ -6,7 +6,7 @@
 use rusqlite::{Connection, Transaction};
 use sql_support::open_database::{self, ConnectionInitializer};
 
-pub const VERSION: u32 = 1;
+pub const VERSION: u32 = 2;
 
 pub const SQL: &str = "
     CREATE TABLE meta(
@@ -26,14 +26,21 @@ pub const SQL: &str = "
     CREATE TABLE suggestions(
         id INTEGER PRIMARY KEY,
         record_id TEXT NOT NULL,
-        block_id INTEGER NOT NULL,
-        advertiser TEXT NOT NULL,
-        iab_category TEXT NOT NULL,
+        provider INTEGER NOT NULL,
         title TEXT NOT NULL,
         url TEXT NOT NULL,
-        icon_id TEXT NOT NULL,
-        impression_url TEXT,
-        click_url TEXT
+        icon_id TEXT NOT NULL
+    );
+
+    CREATE TABLE amp_custom_details(
+        suggestion_id INTEGER PRIMARY KEY,
+        advertiser TEXT NOT NULL,
+        block_id INTEGER NOT NULL,
+        iab_category TEXT NOT NULL,
+        impression_url TEXT NOT NULL,
+        click_url TEXT NOT NULL,
+        FOREIGN KEY(suggestion_id) REFERENCES suggestions(id)
+        ON DELETE CASCADE
     );
 
     CREATE INDEX suggestions_record_id ON suggestions(record_id);